// src/interfaces/commonInterfaces.ts
export interface UserMode {
  id: number;
  deptId: number;
  loginName: string;
  userName: string;
  userType: string;
  email: string;
  phonenumber: number;
  sex: string;
  avatar: string;
  status: string;
  delFlag: string;
  loginIp: string;
  loginDate: Date;
  createBy: string;
  createTime: string;
}